+++ server/index.js
@@ -4,7 +4,7 @@ const http = require('http').Server(app);
 const io = require('socket.io')(http);
 
 const { game } = require('./game');
-const { player } = require('./player');
+const { Player } = require('./player');
 const { greetSinglePlayer, waveOtherPlayers, messageAllPlayers } = require('./message-players');
 const { spawnMultipleNpc } = require('./spawn-multiple-npc');
 const { spawnPlayer, updatePlayerState  } = require('./spawn-players');
@@ -39,7 +39,7 @@ io.on('connection', (socket) => {
     }
 
     // Create the new player.
-    dataStore.players[name] = player(name);
+    dataStore.players[name] = new Player({ name });
 
     // Messages
     greetSinglePlayer(io, socket.id, name);