+++ server/player.js
@@ -1,21 +1,19 @@
+const { Wisp } = require('./wisp');
 const { deepCopy } = require('./utils');
-const { createWisp } = require('./create-wisp');
 
-const player = (name) => {
-  const player = {
-    creation: createWisp({
-      name,
-      position: [0, 0, 0]
-    })
-  };
+class Player extends Wisp {
+  constructor(input) {
+    super(input);
 
-  player.creation.color[0] = 0.3583; // Hue.
+    this.position = [0, 0, 0];
+    this.color[0] = 0.3583; // Hue.
 
-  player.state = deepCopy(player.creation);
+    this.creation = deepCopy(this);
 
-  return player;
-};
+    this.state = deepCopy(this.creation);
+  }
+}
 
 module.exports = {
-  player,
+  Player,
 };